% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myCRAN.R
\name{myCRAN}
\alias{myCRAN}
\title{myCRAN}
\usage{
myCRAN(
  packages = NULL,
  author = NULL,
  from,
  when,
  plotNew = TRUE,
  plotWidth = 12
)
}
\arguments{
\item{packages}{character vector of names of packages}

\item{author}{character string of name of target author}

\item{from}{date parameter for cran_downloads()}

\item{when}{date parameter for cran_downloads()}

\item{plotNew}{Boolean if TRUE start a new plot window}

\item{plotWidth}{integer width parameter for dev.new()}
}
\value{
returns no values, but has side effect of generating plots
of daily and cumulative number of downloads of your packages
}
\description{
wrapper for cran_downloads() to plot counts and cumulative counts
of downloads of CRAN packages
}
\details{
Use either packages or author as the input parameter.
When author is given, the function advanced_search() queries CRAN
to return a list of the names of all packages by that author.
}
\examples{
\dontrun{
	packages<-c("timeLineGraphics","textBoxPlacement","SherlockHolmes","myCRAN")
	author = "Zeeberg"
	from<-"2023-01-01"
	myCRAN(author=author,from=from,plotNew=TRUE)
	myCRAN(packages=packages,when="last-week",plotNew=FALSE)
	}

}
