% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myCRAN.R
\name{myCRAN}
\alias{myCRAN}
\title{myCRAN}
\usage{
myCRAN(packages, from, when, plotNew = TRUE, plotWidth = 15)
}
\arguments{
\item{packages}{character vector of names of packages}

\item{from}{date parameter for cran_downloads()}

\item{when}{date parameter for cran_downloads()}

\item{plotNew}{Boolean if TRUE start a new plot window}

\item{plotWidth}{integer width parameter for dev.new()}
}
\value{
returns no values, but has side effect of generating plots
of daily and cumulative number of downloads of your packages
}
\description{
wrapper for cran_downloads() to plot counts and cumulative counts
of downloads of CRAN packages
}
\examples{
	packages<-c("timeLineGraphics","textBoxPlacement","SherlockHolmes")
	from<-"2023-01-01"
	myCRAN(packages,from=from,plotNew=TRUE)
	myCRAN(packages,when="last-week",plotNew=FALSE)

}
